backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

pydantic==2.8.2
python-multipart==0.0.9

sqlalchemy==2.0.34
passlib[bcrypt]==1.7.4
python-jose==3.3.0

sentence-transformers==3.0.1
torch==2.4.1
transformers==4.44.2

numpy==1.26.4
scikit-learn==1.5.1
scipy==1.14.1

requests==2.32.3
httpx==0.27.2

python-dotenv==1.0.1
pip==26.0.1
faiss-cpu==1.8.0
openpyxl
easyocr==1.7.1
opencv-python==4.10.0.84
pymupdf==1.24.11
